Composers refactor
+ Creates a composer function inside Actions
+ Refactors the List, Show, Create and Edit Composers
